==> pages/4_0. Previsão_da_produtividade.py <==
import streamlit as st
st.set_page_config(page_title="Previsão da produtividade", layout="wide")
st.markdown("# Previsão da produtividade")
st.page_link("pages/4.1_Análise_de_correlação.py", label="Análise de correlação")
st.page_link("pages/4.2_Treinamento_com_Machine_Learning.py", label="Treinamento com Machine Learning")
st.page_link("pages/4.3_Previsão_da_safra.py", label="Previsão da safra")
